+++ CHANGELOG.md
@@ -15,7 +15,7 @@
 - Truncate text in left & right panels cleanly with '..', avoiding text overflow
 - Show user status as `active` (Green), `idle` (Yellow) or `offline` (White) using different colors.
 - Further improvement/reordering of shortcut keys in README & help menu (<kbd>?</kbd>)
-- Improve styling of help menu, and how the menu scales with application width
+- Improve styling of help menu, and how the menu scales with application width & height
 - Make stream icons bold and correct background color
 
 ### Important bugfixes